requirements.txt
pymongo[srv]==4.7.3
elasticsearch==8.13.1
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
loguru==0.7.2
requests==2.32.3
numpy==1.26.4